requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==4.2.2
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
